P_2_4_1_01_multiuser_canvas PR
This is a PR for a new sketch which uses web sockets and node.js to
connect different browser windows together.

NOTE: I couldn’t think of a good way to serve the necessary libraries
to the sketch besides putting them in the “public” folder, so I’m happy
to discuss solutions!