P_2_2_6_01: Added simplified pendulum sketch
@Niels-NTG : I’ve done quite a lot of renaming of the variables and
restructuring for this version of the sketch. I changed some of the
calculations to make them easier to explain/understand. Also, I removed
the mouse drawing functionality, because all the drawing examples
happen a little later in the book.
There is still some potential to make things better (I’m not happy with
all these `new Shape()` calls.
Could you please have a quick look if these changes are ok with you?