Added P_3_1_4_01 as an approach to rebuild the old sketch
As @Niels-NTG mentioned in #76, I agree that it quite hard/impossible
to imitate the old sketch exactly. I’ve written a class, that at least
manages to draw one kind of treemap. Still a bit messy and to hard to
explain in the book. I’ll try to make this one recursive. Then it might
be possible to do the word and word type counting as in P_3_1_3 and
draw it as a nested treemap.